rename: Banner 컴포넌트 폴더명 수정
